@@ -37,79 +40,53 @@
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
     "@mui/material": "^5.11.11",
+    "gas-client": "^1.1.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
-    "react-transition-group": "^4.4.2",
-    "tailwindcss": "^3.1.6"
+    "react-transition-group": "^4.4.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.6",
-    "@babel/core": "^7.17.8",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.17.12",
-    "@babel/plugin-proposal-object-rest-spread": "^7.18.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.17.12",
-    "@babel/plugin-transform-object-assign": "^7.16.7",
-    "@babel/polyfill": "^7.12.1",
-    "@babel/preset-env": "^7.18.2",
-    "@babel/preset-react": "^7.17.12",
-    "@effortlessmotion/dynamic-cdn-webpack-plugin": "^5.0.1",
-    "@effortlessmotion/html-webpack-inline-source-plugin": "^1.0.3",
-    "@google/clasp": "^2.4.1",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@babel/core": "^7.24.4",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.24.4",
+    "@google/clasp": "^2.4.2",
     "@types/expect-puppeteer": "^5.0.0",
     "@types/jest-environment-puppeteer": "^5.0.2",
+    "@types/node": "^20.11.30",
     "@types/puppeteer": "^5.4.6",
-    "@types/react": "^18.0.14",
-    "autoprefixer": "10.4.5",
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "autoprefixer": "^10.4.19",
     "aws-sdk": "^2.1106.0",
-    "babel-loader": "^8.2.5",
-    "babel-plugin-add-module-exports": "^1.0.4",
-    "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
-    "babel-plugin-transform-es3-property-literals": "^6.22.0",
-    "copy-webpack-plugin": "^11.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
-    "dotenv": "^16.0.1",
-    "eslint": "^8.17.0",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-googleappsscript": "^1.0.4",
-    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^26.5.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.30.0",
-    "eslint-plugin-standard": "^5.0.0",
-    "gas-client": "^1.1.1",
-    "gas-lib": "^2.0.4",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.6",
     "gas-types-detailed": "^1.1.2",
-    "gas-webpack-plugin": "^2.2.2",
-    "html-webpack-plugin": "^5.5.0",
     "jest": "^28.1.1",
     "jest-environment-node": "^28.1.1",
     "jest-image-snapshot": "^5.1.0",
-    "mkdirp": "^1.0.4",
-    "module-to-cdn": "^3.1.5",
-    "postcss-loader": "^7.0.1",
-    "postcss-preset-env": "^7.7.2",
+    "postcss": "^8.4.38",
+    "postcss-preset-env": "^9.5.4",
     "prettier": "^2.7.0",
     "puppeteer": "^14.3.0",
     "puppeteer-extra": "^3.2.3",
     "puppeteer-extra-plugin-stealth": "^2.9.0",
-    "react-refresh": "^0.14.0",
-    "rimraf": "^3.0.2",
-    "style-loader": "^3.3.1",
-    "terser-webpack-plugin": "^5.3.3",
-    "ts-loader": "^9.3.0",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.9.2"
+    "tailwindcss": "^3.4.3",
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0",
+    "vite-plugin-singlefile": "^2.0.1",
+    "vite-plugin-static-copy": "^1.0.1"
   }
 }
