@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   globalSetup: './test/global-setup.js',
   globalTeardown: './test/global-teardown.js',
   testEnvironment: './test/puppeteer-environment.js',
