@@ -2,15 +2,16 @@
 // jestjs.io/docs/puppeteer#custom-example-without-jest-puppeteer-preset
 // This allows using stealth mode.
 
-const fs = require('fs').promises;
-const os = require('os');
-const path = require('path');
+import fs from 'fs';
+const fsPromises = fs.promises;
+import os from 'os';
+import path from 'path';
 
 const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
-module.exports = async function globalTeardown() {
+export default async function globalTeardown() {
   // close the browser instance
   await global.__BROWSER_GLOBAL__.close();
 
   // clean-up the wsEndpoint file
-  await fs.rmdir(DIR, { recursive: true, force: true });
+  await fsPromises.rmdir(DIR, { recursive: true, force: true });
 };
