@@ -2,15 +2,15 @@
 // jestjs.io/docs/puppeteer#custom-example-without-jest-puppeteer-preset
 // This allows using stealth mode.
 
-const { readFile } = require('fs').promises;
-const os = require('os');
-const path = require('path');
-const puppeteer = require('puppeteer');
-const NodeEnvironment = require('jest-environment-node').default;
+import { readFile } from 'fs/promises';
+import os from 'os';
+import path from 'path';
+import puppeteer from 'puppeteer';
+import NodeEnvironment from 'jest-environment-node';
 
 const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
 
-class PuppeteerEnvironment extends NodeEnvironment {
+export default class PuppeteerEnvironment extends NodeEnvironment.default {
   constructor(config) {
     super(config);
   }
@@ -37,5 +37,3 @@ class PuppeteerEnvironment extends NodeEnvironment {
     return super.getVmContext();
   }
 }
-
-module.exports = PuppeteerEnvironment;
